eman updated courses experiences 25-1-2023

+++ database/migrations/2023_01_16_115309_create_courses_table.php
@@ -16,11 +16,14 @@ class CreateCoursesTable extends Migration
         Schema::create('courses', function (Blueprint $table) {
             $table->id();
             $table->string('name',255);
-            $table->mediumInteger('hours');
+            $table->integer('hours');
             $table->string('training_center',255);
             $table->date('end_date');
             $table->boolean('status')->default(1);
-            $table->foreignId('user_id')->references('id')->on('users')->onDelete('cascade');
+            $table->foreignId('user_id')
+                ->references('id')
+                ->on('users')
+                ->onDelete('cascade');
             $table->timestamps();
         });
     }
